Restructure utility methods to make consistent pattern
