+++ packages/prover/src/verified_requests/eth_getBalance.ts
@@ -1,5 +1,5 @@
 import {ELVerifiedRequestHandler} from "../interfaces.js";
-import {fetchAndVerifyAccount} from "../utils/execution.js";
+import {verifyAccount} from "../utils/verification.js";
 import {generateRPCResponseForPayload, generateUnverifiedResponseForPayload} from "../utils/json_rpc.js";
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
@@ -12,7 +12,7 @@ export const eth_getBalance: ELVerifiedRequestHandler<[address: string, block?: 
   const {
     params: [address, block],
   } = payload;
-  const result = await fetchAndVerifyAccount({proofProvider, logger, handler, address, block});
+  const result = await verifyAccount({proofProvider, logger, handler, address, block});
 
   if (result.valid) {
     return generateRPCResponseForPayload(payload, result.data.balance);
