+++ packages/prover/src/verified_requests/eth_getBlockByHash.ts
@@ -1,7 +1,7 @@
 import {ELVerifiedRequestHandler} from "../interfaces.js";
 import {ELBlock} from "../types.js";
-import {fetchAndVerifyBlock} from "../utils/execution.js";
-import {generateUnverifiedResponseForPayload} from "../utils/json_rpc.js";
+import {verifyBlock} from "../utils/verification.js";
+import {generateUnverifiedResponseForPayload, generateVerifiedResponseForPayload} from "../utils/json_rpc.js";
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export const eth_getBlockByHash: ELVerifiedRequestHandler<[block: string, hydrated: boolean], ELBlock> = async ({
@@ -11,10 +11,10 @@ export const eth_getBlockByHash: ELVerifiedRequestHandler<[block: string, hydrat
   proofProvider,
   network,
 }) => {
-  const result = await fetchAndVerifyBlock({payload, proofProvider, logger, handler, network});
+  const result = await verifyBlock({payload, proofProvider, logger, handler, network});
 
   if (result.valid) {
-    return result.data;
+    return generateVerifiedResponseForPayload(payload, result.data);
   }
 
   logger.error("Request could not be verified.");
