+++ packages/prover/src/verified_requests/eth_getCode.ts
@@ -1,5 +1,5 @@
 import {ELVerifiedRequestHandler} from "../interfaces.js";
-import {fetchAndVerifyAccount, fetchAndVerifyCode} from "../utils/execution.js";
+import {verifyAccount, verifyCode} from "../utils/verification.js";
 import {generateRPCResponseForPayload, generateUnverifiedResponseForPayload} from "../utils/json_rpc.js";
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
@@ -13,7 +13,7 @@ export const eth_getCode: ELVerifiedRequestHandler<[address: string, block?: num
     params: [address, block],
   } = payload;
   // TODO: When batch requests are supported merged these two requests into one
-  const accountProof = await fetchAndVerifyAccount({
+  const accountProof = await verifyAccount({
     proofProvider,
     logger,
     handler,
@@ -26,7 +26,7 @@ export const eth_getCode: ELVerifiedRequestHandler<[address: string, block?: num
     return generateUnverifiedResponseForPayload(payload, "account for eth_getCode request can not be verified.");
   }
 
-  const codeProof = await fetchAndVerifyCode({
+  const codeProof = await verifyCode({
     proofProvider,
     logger,
     handler,
